live-cta/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "live-cta"
version = "0.1.0"
description = "Live inference pipeline for CTAFlow CTA models"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "torch>=1.10.0",
]

[project.optional-dependencies]
ibkr = ["requests>=2.25.0"]
s3 = ["boto3>=1.20.0"]
gcs = ["google-cloud-storage>=2.0.0"]
dashboard = ["pyarrow>=6.0.0"]
server = ["fastapi>=0.68.0", "uvicorn>=0.15.0"]
all = ["live-cta[ibkr,s3,gcs,dashboard,server]"]
dev = ["pytest>=6.0", "pytest-cov>=2.10"]

[project.scripts]
ng-inference = "live_cta.scripts.run_ng_inference:main"
ng-feeder = "live_cta.scripts.local_feeder:main"
ng-server = "live_cta.scripts.inference_server:main"

[tool.setuptools]
packages = {find = {include = ["live_cta*"], exclude = ["tests*"]}}
